restructure the server code into separate files
