+++ src/server/persistent-resource.js
